Python/requirements.txt
opencv-python
apriltag
cmake
pynetworktables
pkgconfig